feat: 修改获取直播地址的接口

+++ packages/48tools/src/pages/Bilibili/Live/function/bilibiliAutoRecord.ts
@@ -1,11 +1,11 @@
 import * as path from 'node:path';
 import type { Store } from '@reduxjs/toolkit';
-import { requestRoomInitData, requestRoomPlayerUrl, type RoomInit, type RoomPlayUrl } from '@48tools-api/bilibili/live';
+import { requestRoomInitData, requestRoomPlayerUrlV2, type RoomInit, type RoomPlayUrlV2 } from '@48tools-api/bilibili/live';
 import { store } from '../../../../store/store';
 import { getFFmpeg, getFileTime } from '../../../../utils/utils';
 import { liveSlice, setAddWorkerItem, setRemoveWorkerItem } from '../../reducers/bilibiliLive';
 import getFFmpegDownloadWorker from '../../../../utils/worker/FFmpegDownload.worker/getFFmpegDownloadWorker';
-import { ffmpegHeaders, localStorageKey } from './helper';
+import { localStorageKey, createV2LiveUrl } from './helper';
 import type { WebWorkerChildItem, MessageEventData } from '../../../../commonTypes';
 import type { LiveSliceInitialState } from '../../../../store/slice/LiveSlice';
 
@@ -18,8 +18,7 @@ async function bilibiliAutoRecord(): Promise<void> {
   for (const record of liveList) {
     if (!record.autoRecord) continue;
 
-    const index: number = liveSlice.getWorkerList(getState().bilibiliLive).findIndex(
-      (o: WebWorkerChildItem): boolean => o.id === record.id);
+    const index: number = liveSlice.getWorkerList(getState().bilibiliLive).findIndex((o: WebWorkerChildItem): boolean => o.id === record.id);
 
     if (index >= 0) continue;
 
@@ -29,35 +28,37 @@ async function bilibiliAutoRecord(): Promise<void> {
       const resInit: RoomInit = await requestRoomInitData(record.roomId);
 
       if (resInit.data.live_status === 1) {
-        const resPlayUrl: RoomPlayUrl = await requestRoomPlayerUrl(`${ resInit.data.room_id }`);
-        const worker: Worker = getFFmpegDownloadWorker();
+        const resPlayUrl: RoomPlayUrlV2 = await requestRoomPlayerUrlV2(`${ resInit.data.room_id }`);
+        const playStreamPath: string | null = createV2LiveUrl(resPlayUrl);
 
-        worker.addEventListener('message', function(messageEvent: MessageEvent<MessageEventData>) {
-          const { type, error }: MessageEventData = messageEvent.data;
+        if (playStreamPath) {
+          const worker: Worker = getFFmpegDownloadWorker();
 
-          if (type === 'close' || type === 'error') {
-            if (type === 'error') {
-              console.error(error);
+          worker.addEventListener('message', function(messageEvent: MessageEvent<MessageEventData>) {
+            const { type, error }: MessageEventData = messageEvent.data;
+
+            if (type === 'close' || type === 'error') {
+              if (type === 'error') {
+                console.error(error);
+              }
+
+              worker.terminate();
+              dispatch(setRemoveWorkerItem(record.id));
             }
+          }, false);
 
-            worker.terminate();
-            dispatch(setRemoveWorkerItem(record.id));
-          }
-        }, false);
+          worker.postMessage({
+            type: 'start',
+            playStreamPath,
+            filePath: path.join(bilibiliAutoRecordSavePath, `${ record.roomId }_${ record.description }_${ time }.flv`),
+            ffmpeg: getFFmpeg()
+          });
 
-        worker.postMessage({
-          type: 'start',
-          playStreamPath: resPlayUrl.data.durl[0].url,
-          filePath: path.join(bilibiliAutoRecordSavePath, `${ record.roomId }_${ record.description }_${ time }.flv`),
-          ffmpeg: getFFmpeg(),
-          ua: true,
-          ffmpegHeaders: ffmpegHeaders()
-        });
-
-        dispatch(setAddWorkerItem({
-          id: record.id,
-          worker
-        }));
+          dispatch(setAddWorkerItem({
+            id: record.id,
+            worker
+          }));
+        }
       }
     } catch (err) {
       console.error(err);
